requirements.txt
# Security Feed Automation Dependencies

# RSS 파싱
feedparser>=6.0.0

# HTTP 요청
requests>=2.31.0

# 설정 파일
pyyaml>=6.0.0

# Google AI (Gemma)
google-genai>=1.0.0

# Slack 알림 (선택, requests로도 충분)
# slack-sdk>=3.21.0
